requirements.txt
dnspython
bottle
pymongo
gunicorn
redis
python-dotenv==0.14.0
